Move errors into their own file
